scripts/requirements_test_server.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
rclpy>=4.0.0
paho-mqtt>=1.6.0
requests>=2.31.0
httpx>=0.24.0
numpy>=1.24.0
psutil>=5.9.0

